requirements.txt
annotated-types==0.5.0
blinker==1.6.2
click==8.1.6
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
psycopg2==2.9.7
pydantic==2.2.1
pydantic-settings==2.0.3
pydantic_core==2.6.1
pydentic==0.0.1.dev3
python-dotenv==1.0.0
python-stdnum==1.19
six==1.16.0
SQLAlchemy==2.0.19
typing_extensions==4.7.1
Werkzeug==2.3.6
